backend/requirements.txt
asgiref==3.8.1
Django==5.2
sqlparse==0.5.3
typing_extensions==4.13.2
googlemaps>=4.10.0
tqdm>=4.64.0
python-dotenv==1.1.0
